# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Central Studio & Stage Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instructors:
    get:
      operationId: listInstructors
      tags: [instructors]
      summary: List all instructors
      responses:
        "200":
          description: List of instructors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instructor"
    post:
      operationId: createInstructor
      tags: [instructors]
      summary: Create instructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstructorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
  /instructors/{id}:
    get:
      operationId: getInstructor
      tags: [instructors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Instructor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInstructor
      tags: [instructors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstructorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInstructor
      tags: [instructors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClass
      tags: [classes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClass
      tags: [classes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: studentEmail
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInBody"
      responses:
        "201":
          description: Attendance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/stats:
    get:
      operationId: getAttendanceStats
      tags: [attendance]
      summary: Get attendance statistics
      parameters:
        - name: period
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Attendance stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStats"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Instructor:
      type: object
      required: [id, name, specialties, experienceYears, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        bio: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        specialties: { type: array, items: { type: string } }
        experienceYears: { type: integer }
        rating: { type: ["number", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    CreateInstructorBody:
      type: object
      required: [name, specialties, experienceYears]
      properties:
        name: { type: string }
        bio: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        specialties: { type: array, items: { type: string } }
        experienceYears: { type: integer }
        rating: { type: ["number", "null"] }
        isActive: { type: boolean }
    UpdateInstructorBody:
      type: object
      properties:
        name: { type: string }
        bio: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        specialties: { type: array, items: { type: string } }
        experienceYears: { type: integer }
        rating: { type: ["number", "null"] }
        isActive: { type: boolean }
    Class:
      type: object
      required: [id, title, category, level, durationMins, capacity, isActive, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        instructorId: { type: ["integer", "null"] }
        category: { type: string }
        level: { type: string }
        durationMins: { type: integer }
        capacity: { type: integer }
        photoUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    CreateClassBody:
      type: object
      required: [title, category, level, durationMins, capacity]
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        instructorId: { type: ["integer", "null"] }
        category: { type: string }
        level: { type: string }
        durationMins: { type: integer }
        capacity: { type: integer }
        photoUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
    UpdateClassBody:
      type: object
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        instructorId: { type: ["integer", "null"] }
        category: { type: string }
        level: { type: string }
        durationMins: { type: integer }
        capacity: { type: integer }
        photoUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
...
    Booking:
      type: object
      required: [id, studentName, studentEmail, status, bookedAt, createdAt]
      properties:
        id: { type: integer }
        studentName: { type: string }
        studentEmail: { type: string }
        studentPhone: { type: ["string", "null"] }
        scheduleId: { type: ["integer", "null"] }
        classId: { type: ["integer", "null"] }
        packageId: { type: ["integer", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        bookedAt: { type: string }
        createdAt: { type: string }
    CreateBookingBody:
      type: object
      required: [studentName, studentEmail]
      properties:
        studentName: { type: string }
        studentEmail: { type: string }
        studentPhone: { type: ["string", "null"] }
        scheduleId: { type: ["integer", "null"] }
        classId: { type: ["integer", "null"] }
        packageId: { type: ["integer", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        bookedAt: { type: string }
    UpdateBookingBody:
      type: object
      properties:
        studentName: { type: string }
        studentEmail: { type: string }
        studentPhone: { type: ["string", "null"] }
        scheduleId: { type: ["integer", "null"] }
        classId: { type: ["integer", "null"] }
        packageId: { type: ["integer", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
...
    Attendance:
      type: object
      required: [id, studentName, studentEmail, creditDeducted, checkedInAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        studentName: { type: string }
        studentEmail: { type: string }
        packageOrderId: { type: ["integer", "null"] }
        classTitle: { type: ["string", "null"] }
        creditDeducted: { type: boolean }
        notes: { type: ["string", "null"] }
        checkedInAt: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    CheckInBody:
      type: object
      required: [studentEmail, studentName]
      properties:
        studentEmail: { type: string }
        studentName: { type: string }
        packageOrderId: { type: ["integer", "null"] }
        classTitle: { type: ["string", "null"] }
        creditDeducted: { type: boolean }
        notes: { type: ["string", "null"] }
    AttendanceStats:
      type: object
      required: [period, total, data]
      properties:
        period: { type: string }
        total: { type: integer }
        data:
          type: array
          items:
            type: object
            required: [label, count]
            properties:
              label: { type: string }
              count: { type: integer }